More memory efficient way to go through the tree
Instead of doing the counting (and more importantly the hashmap) of the full
tree, do it in pieces.  The way to do the pieces is to use captures as
irreversible moves I.e. first explore the full tree as now, but don't follow
captures. Instead, store the positions after the capture in a map (with their
required depth), for the next round of iterations. So, instead of having to
store all positions, you end up storing just those with N and N-1 pieces.
Gives nice additional stats on the number of positions at a given depth as a
function of piece count as well.  With some effort pawn moves could be put into
the picture as well, but that's a bit more tricky I think.  (The map with N-1
pieces is smaller as well, since that's just those immediately after a
capture).